revert gate numerics, have the dma async back

+++ tpu_inference/kernels/gdn/v2/recurrent_scan_v2.py
@@ -452,7 +452,10 @@ def inner_kernel(
             a_raw_processed = a_raw_chunk[:, :n_v].T
             b_raw_processed = b_raw_chunk[:, :n_v].T
 
-            # Compute gates in VMEM in full fp32, not sure if needed.
+            # Compute gates in VMEM in full fp32. Casting beta input to fp32
+            # so beta = sigmoid(fp32) instead of sigmoid(bf16) (baseline numerics).
+            a_raw_processed = a_raw_processed.astype(jnp.float32)
+            b_raw_processed = b_raw_processed.astype(jnp.float32)
             beta = jax.nn.sigmoid(b_raw_processed)
             g = -jnp.exp(a_log_ref[...][:, None].astype(
                 jnp.float32)) * jax.nn.softplus(a_raw_processed + dt_bias_ref[
@@ -634,9 +637,11 @@ def inner_kernel(
             a_raw_processed = a_raw_chunk[:C_trans, :n_v].T
             b_raw_processed = b_raw_chunk[:C_trans, :n_v].T
 
-            # NOTE: b is upcasted to f32 in ref before sigmoid, beta is bf16
+            # Cast gate inputs to fp32 so beta_chunk = sigmoid(fp32) instead of
+            # sigmoid(bf16) (baseline numerics).
+            a_raw_processed = a_raw_processed.astype(jnp.float32)
+            b_raw_processed = b_raw_processed.astype(jnp.float32)
             beta_chunk = jax.nn.sigmoid(b_raw_processed)
-            # NOTE: a is upcasted to f32 before add to dt_bias
             g_chunk = -jnp.exp(a_log_ref[...][:, None].astype(
                 jnp.float32)) * jax.nn.softplus(a_raw_processed + dt_bias_ref[
                     ...][:, None].astype(jnp.float32))
